Generate System Diagram
